Factor out falcon_oas.oas validation as oas library
