+++ setup.cfg
@@ -20,15 +20,10 @@ package_dir =
     = src
 install_requires =
     falcon >= 1.4, < 3
-    jsonref ~= 0.2
-    jsonschema ~= 2.6
+    oas ~= 0.1.0
     six ~= 1.11
-    functools32 ~= 3.2; python_version == '2.7'
 
 [options.extras_require]
-format =
-    pyrfc3339 ~= 1.1
-    rfc3986 ~= 1.1
 test =
     pytest ~= 4.4
     pytest-cov ~= 2.6
