+++ tests/test_middlewares.py
@@ -7,15 +7,15 @@ from __future__ import unicode_literals
 import falcon
 import pytest
 from falcon import testing
+from oas import create_spec_from_dict
+from oas.exceptions import UndocumentedMediaType
+from oas.exceptions import UnmarshalError
 
 import falcon_oas
 from falcon_oas import extensions
+from falcon_oas.exceptions import SecurityError
 from falcon_oas.middlewares import _get_security_schemes
 from falcon_oas.middlewares import _RequestAdapter
-from falcon_oas.oas.exceptions import SecurityError
-from falcon_oas.oas.exceptions import UndocumentedMediaType
-from falcon_oas.oas.exceptions import UnmarshalError
-from falcon_oas.oas.spec import create_spec_from_dict
 
 
 USER = object()
