+++ tests/test_routing.py
@@ -4,8 +4,9 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from oas import create_spec_from_dict
+
 from falcon_oas import extensions
-from falcon_oas.oas.spec import create_spec_from_dict
 from falcon_oas.routing import generate_routes
