+++ tox.ini
@@ -7,7 +7,6 @@ deps =
     falcon1: falcon ~= 1.4
     falcon2: falcon ~= 2.0
 extras =
-    format
     test
 commands =
     pytest --cov --cov-report=term {posargs}
